third_party/WebKit/LayoutTests/fast/forms/select/select-size.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<select></select>
<script>
description('Test HTMLSelectElement::size behavior.');
var select = document.querySelector('select');
shouldBe('select.size', '0');
shouldBe('select.setAttribute("size", "1"); select.size', '1');
shouldBe('select.setAttribute("size", "2"); select.size', '2');
shouldBe('select.setAttribute("size", "3"); select.size', '3');
shouldBe('select.setAttribute("size", "4"); select.size', '4');
shouldBe('select.setAttribute("size", "0"); select.size', '0');
shouldBe('select.setAttribute("size", "-1"); select.size', '0');
shouldBe('select.setAttribute("size", "abc"); select.size', '0');
shouldBe('select.setAttribute("size", "3.14"); select.size', '3');
</script>
