third_party/blink/web_tests/platform/generic/fast/forms/select/select-size-expected.txt

Test HTMLSelectElement::size behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS select.size is 0
PASS select.setAttribute("size", "1"); select.size is 1
PASS select.setAttribute("size", "2"); select.size is 2
PASS select.setAttribute("size", "3"); select.size is 3
PASS select.setAttribute("size", "4"); select.size is 4
PASS select.setAttribute("size", "0"); select.size is 0
PASS select.setAttribute("size", "-1"); select.size is 0
PASS select.setAttribute("size", "abc"); select.size is 0
PASS select.setAttribute("size", "3.14"); select.size is 3
PASS successfullyParsed is true

TEST COMPLETE
